+++ graylog2-server/src/main/java/org/graylog/plugins/views/ViewsModule.java
@@ -27,8 +27,8 @@ import org.graylog.plugins.views.search.Search;
 import org.graylog.plugins.views.search.engine.GeneratedQueryContext;
 import org.graylog.plugins.views.search.engine.QueryBackend;
 import org.graylog.plugins.views.search.engine.QueryStringDecorator;
-import org.graylog.plugins.views.search.engine.SearchNormalizer;
-import org.graylog.plugins.views.search.engine.SearchValidator;
+import org.graylog.plugins.views.search.engine.normalization.SearchNormalizer;
+import org.graylog.plugins.views.search.engine.validation.SearchValidator;
 import org.graylog.plugins.views.search.export.ExportBackend;
 import org.graylog.plugins.views.search.rest.SeriesDescription;
 import org.graylog.plugins.views.search.searchtypes.pivot.SeriesSpec;
