+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/engine/normalization/SearchNormalization.java
@@ -1,9 +1,9 @@
-package org.graylog.plugins.views.search.engine;
+package org.graylog.plugins.views.search.engine.normalization;
 
 import org.graylog.plugins.views.search.Search;
 import org.graylog.plugins.views.search.permissions.SearchUser;
 import org.graylog.plugins.views.search.rest.ExecutionState;
 
-public interface SearchNormalizer {
+public interface SearchNormalization {
     Search normalize(Search search, SearchUser searchUser, ExecutionState executionState);
 }
