+++ graylog2-server/src/main/java/org/graylog/plugins/views/search/engine/validation/PluggableSearchValidation.java
@@ -1,4 +1,4 @@
-package org.graylog.plugins.views.search.engine;
+package org.graylog.plugins.views.search.engine.validation;
 
 import org.graylog.plugins.views.search.Search;
 import org.graylog.plugins.views.search.SearchExecutionGuard;
